+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsEventParams.kt
@@ -8,7 +8,6 @@ import androidx.annotation.RestrictTo
  * all parameters are required at each call site.
  *
  * @property payPalContextId Used for linking events from the client to server side request.
- * @property linkType Indicates whether a deeplink or an app link was used to launch the app. Also see [LinkType].
  * @property isVaultRequest Indicates whether the request was a BillingAgreement(BA)/Vault request.
  * @property startTime [HttpResponseTiming] start time.
  * @property endTime [HttpResponseTiming] end time.
@@ -20,11 +19,11 @@ import androidx.annotation.RestrictTo
  * @property buttonType buttonType Represents the tapped button type.
  * @property buttonOrder The order or ranking in which payment buttons appear.
  * @property pageType The page or view that a button is displayed on.
+ * @property errorDescription use this to provide more detail on a failure event.
  */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class AnalyticsEventParams @JvmOverloads constructor(
     val payPalContextId: String? = null,
-    val linkType: String? = null,
     val isVaultRequest: Boolean = false,
     val startTime: Long? = null,
     val endTime: Long? = null,
@@ -34,5 +33,6 @@ data class AnalyticsEventParams @JvmOverloads constructor(
     val shopperSessionId: String? = null,
     val buttonType: String? = null,
     val buttonOrder: String? = null,
-    val pageType: String? = null
+    val pageType: String? = null,
+    val errorDescription: String? = null,
 )